chore(wallet): refactor wallet module
